requirements.txt
sqlalchemy~=2.0.44
bcrypt~=5.0.0
PyJWT~=2.10.1
uvicorn~=0.38.0
fastapi~=0.123.4
pydantic~=2.12.5
python-multipart~=0.0.20